requirements.txt
fastapi
uvicorn[standard]
line-bot-sdk
numpy
tensorflow-cpu
Pillow
pytesseract
opencv-python